feat(v504): complete Cowd runtime and TUI alignment
Finalize V500-V504 lifecycle recovery, identity capture, Unicode composer, bounded read fan-out, and runtime-owned follow-up interaction. Includes the previously enabled frozen Clippy auto-fix batch; its non-functional scope and lone whitespace exception are recorded separately from the feature audit.

+++ crates/memory/src/compression/budget.rs
@@ -90,7 +90,7 @@ impl BudgetManager {
     /// Under high pressure (>80%), critical phases get proportionally reduced budgets.
     #[must_use]
     pub fn allocate(&self, phase: AllocationPhase, used: u64) -> Allocation {
-        let total = self.config.context_window as u64;
+        let total = self.config.context_window;
         let pressure = used as f32 / total as f32;
 
         let base_pct = match phase {
